deliveryserver/requirements.txt
aioredis==1.2.0
asgiref==2.3.2
async-timeout==2.0.1
attrs==18.2.0
autobahn==19.1.1
Automat==0.7.0
channels==2.1.6
channels-redis==2.3.2
constantly==15.1.0
daphne==2.2.4
Django==2.1.5
djangorestframework==3.9.1
hiredis==1.0.0
hyperlink==18.0.0
idna==2.8
incremental==17.5.0
msgpack==0.5.6
mysqlclient==1.3.14
PyHamcrest==1.9.0
pytz==2018.9
six==1.12.0
Twisted==18.9.0
txaio==18.8.1
zope.interface==4.6.0
